fix: monitor nodes task
